Rename MetaconfigPendingUpstream to MetaconfigOps (fix #499)
Those methods can't easily be upstreamed since some of them are
scalameta/scalafix specific.
